Address route code review requests
