+++ crates/alkahest/src/gui/inspector/mod.rs
@@ -16,9 +16,10 @@ use alkahest_renderer::{
             light::LightRenderer,
         },
         resources::SelectedEntity,
+        route::{Route, RouteNode},
         tags::{insert_tag, remove_tag, EntityTag, Tags},
         transform::{OriginalTransform, Transform, TransformFlags},
-        utility::{Beacon, Route, RouteNode, Ruler, Sphere},
+        utility::{Beacon, Ruler, Sphere},
         visibility::{Visibility, VisibilityHelper},
         Scene,
     },
